minor fixes for compiler warnings
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@6245 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ examples/segmentation/example_extract_clusters_normals.cpp
@@ -98,7 +98,7 @@ main (int, char **argv)
     pcl::PointCloud<pcl::PointXYZ>::Ptr cloud_cluster (new pcl::PointCloud<pcl::PointXYZ>);
     for (std::vector<int>::const_iterator pit = it->indices.begin (); pit != it->indices.end (); pit++)
       cloud_cluster->points.push_back (cloud_ptr->points[*pit]); 
-    cloud_cluster->width = cloud_cluster->points.size ();
+    cloud_cluster->width = static_cast<uint32_t> (cloud_cluster->points.size ());
     cloud_cluster->height = 1;
     cloud_cluster->is_dense = true;
